Interface mais amigável para utilizar o gppanel. Além disso, agora o exemplo plota os gráficos de barra, linha e área juntos.
