+++ gpPanel/gpSeries.cpp
@@ -6,8 +6,10 @@
  */
 gpSeries::gpSeries( wxString label )
 {
-    m_Layer = new lineChartLayer( label );
+    m_lineLayer = new lineChartLayer( label );
 	m_barLayer = new barChartLayer(label);
+	m_areaLayer = new areaChartLayer(label);
+	m_barLayer->SetGradientBackColour(false);
     m_Label = label;
 }
 
@@ -17,12 +19,9 @@ gpSeries::gpSeries( wxString label )
  */
 gpSeries::~gpSeries( void )
 {
-    //delete m_Layer;
-}
-
-void gpSeries::invertY()
-{
-	m_Data.invertY();
+    /*delete m_lineLayer;
+	delete m_barLayer;
+	delete m_areaLayer;*/
 }
 
 
@@ -49,50 +48,9 @@ void gpSeries::RefreshChart( gpCHART_KIND gpChart_kind, double samplerate,
             wxString customYXFormula, gpAXIS_SCALE gpXaxis_type,
             gpAXIS_SCALE gpYaxis_type )
 {
-    m_Layer->DataSet( m_Data.GetData() );
+	m_lineLayer->DataSet( m_Data.GetData() );
 	m_barLayer->DataSet(m_Data.GetData());
-	m_barLayer->SetGradientBackColour(false);
-    if(gpChart_kind== gpCHART_DEFAULT)
-    {
-        m_Layer->DataSet( m_Data.GetData() );
-    }
-
-    else if(gpChart_kind== gpCHART_INL)
-    {
-        m_Layer->DataSet(m_Data.GetINL(true));
-    }
-    else if(gpChart_kind== gpCHART_DNL)
-    {
-        m_Layer->DataSet(m_Data.GetDNL(true));
-    }
-    else if(gpChart_kind== gpCHART_FFT)
-    {
-        m_Layer->DataSet( m_Data.GetFft(fftlenght, samplerate) );
-    }
-    else if(gpChart_kind== gpCHART_CUSTOM)
-    {
-        m_Layer->DataSet( m_Data.GetCustom(wxEmptyString,
-                customYXFormula) );
-    }
-    else if(gpChart_kind== gpCHART_ACCUMULATION)
-    {
-        m_Layer->DataSet( m_Data.GetAccumulation(
-                            gpYaxis_type == gpAXIS_PROCENT ) );
-
-        if(m_Layer->GetMinY()>0)m_Layer->SetMinY(0);
-    }
-
-    if( gpYaxis_type == gpAXIS_CUSTOM )
-    {
-        m_Layer->DataSet( m_Layer->GetCustom(
-                            wxEmptyString, customYFormula) );
-    }
-
-    if( gpXaxis_type == gpAXIS_CUSTOM )
-    {
-        m_Layer->DataSet( m_Layer->GetCustom(
-                    customXFormula, wxEmptyString) );
-    }
+	m_areaLayer->DataSet(m_Data.GetData());
 }
 
 /*!
@@ -122,8 +80,9 @@ bool gpSeries::IsLabel( wxString CompareLabel )
  */
 void gpSeries::SetContinuity( bool continuity )
 {
-    m_Layer->SetContinuity( continuity );
+	m_lineLayer->SetContinuity( continuity );
 	m_barLayer->SetContinuity(continuity);
+	m_areaLayer->SetContinuity(continuity);
 }
 
 
@@ -132,8 +91,9 @@ void gpSeries::SetContinuity( bool continuity )
  */
 void gpSeries::SetVisible( bool show )
 {
-    m_Layer->SetVisible( show );
+	m_lineLayer->SetVisible( show );
 	m_barLayer->SetVisible(show);
+	m_areaLayer->SetVisible(show);
 }
 
 
@@ -142,37 +102,43 @@ void gpSeries::SetVisible( bool show )
  */
 void gpSeries::ShowName( bool show )
 {
-   m_Layer->ShowName( show );
-   m_barLayer->ShowName(show);
+	m_lineLayer->ShowName( show );
+	m_barLayer->ShowName(show);
+	m_areaLayer->ShowName(show);
 }
 
-/*!
- *  \brief Gets a pointer to m_Layer
- *
- *  \return
- *  a pointer to m_Layer
- */
-mpLayer* gpSeries::GetLayer( void )
+
+lineChartLayer* gpSeries::getLineChartLayer( void )
 {
-    return m_Layer;
+    return m_lineLayer;
 }
 
+barChartLayer * gpSeries::getBarChartLayer()
+{
+	return m_barLayer;
+}
+
+areaChartLayer * gpSeries::getAreaChartLayer()
+{
+	return m_areaLayer;
+}
 
 /*!
  *  \brief Calls the SetPen function of m_Layer
  */
 void gpSeries::SetPen( wxPen pen )
 {
-    m_Layer->SetPen( pen );
+	m_lineLayer->SetPen( pen );
 	m_barLayer->SetPen(pen);
+	m_areaLayer->SetPen(pen);
 }
 
 void gpSeries::SetBrush(wxBrush brush)
 {
-	m_Layer->SetBrush(brush);
+	m_lineLayer->SetBrush(brush);
 	m_barLayer->SetBrush(brush);
+	m_areaLayer->SetBrush(brush);
 }
-
 
 /*!
  *  \brief Calls the DataClear function of m_Data
@@ -182,6 +148,12 @@ void gpSeries::DataClear( void )
     m_Data.DataClear();
 }
 
+void gpSeries::invert(bool value)
+{
+	m_lineLayer->invert(value);
+	m_barLayer->invert(value);
+	m_areaLayer->invert(value);
+}
 
 /*!
  *  \brief Gets the multimap data from m_Data
@@ -192,9 +164,4 @@ void gpSeries::DataClear( void )
 xyMultimap_t gpSeries::GetData( void )
 {
     return m_Data.GetData();
-}
-
-barChartLayer * gpSeries::getBarChartLayer()
-{
-	return m_barLayer;
 }
