+++ gpPanel/mathplot.cpp
@@ -120,6 +120,7 @@ mpLayer::mpLayer() : m_type(mpLAYER_UNDEF)
     m_showName   = TRUE;  // Default
     m_drawOutsideMargins = TRUE;
 	m_visible = true;
+	m_invert = false;
 
 	m_pointType  = mpPOINT;
 	m_RectSize = wxSize(4,4);
@@ -551,8 +552,14 @@ void mpFXYBar::Plot(wxDC & dc, mpWindow & w)
                 if(!m_gradienBackground || widthPx==1 || heightPx==1 )
                 {
 					//Precisa fazer checagem se o gráfico é invertido ou não
-                  //dc.DrawRectangle(xPx, yPx, widthPx, heightPx);
-					dc.DrawRectangle(xPx, 0, widthPx, heightPx);
+					if (!m_invert)
+					{
+						dc.DrawRectangle(xPx, yPx, widthPx, heightPx);
+					}
+					else
+					{
+						dc.DrawRectangle(xPx, w.y2p(w.getChartAxisMaxY()), widthPx, heightPx);
+					}
                 }
                 else{
                     dc.GradientFillLinear( wxRect( xPx, yPx, widthPx/2, heightPx),
@@ -564,8 +571,6 @@ void mpFXYBar::Plot(wxDC & dc, mpWindow & w)
                                                   wxColour(150,150,255),
                                                   wxDirection(wxLEFT) );
                 }
-
-
 
                 if( m_showLabel )
                 {
@@ -1531,8 +1536,14 @@ void mpFXY::Plot(wxDC & dc, mpWindow & w)
                                 x1 = endPx;
                             }
                         }
-
-						dc.DrawLine(x0, c0, x1, c1);
+						if (!m_invert)
+						{
+							dc.DrawLine(x0, c0, x1, c1);
+						}
+						else
+						{
+							dc.DrawLine(x0, w.y2p(w.getChartAxisMinY()) - c0, x1, w.y2p(w.getChartAxisMinY()) - c1);
+						}
 						UpdateViewBoundary(x1, c1);
 
 						if(m_markCorners)
@@ -2073,10 +2084,15 @@ void mpScaleY::Plot(wxDC & dc, mpWindow & w)
                     m_pen.SetStyle(wxPENSTYLE_SOLID);
                     dc.SetPen( m_pen);
                 }
-                // Print ticks labels
-                //s.Printf(fmt, std::fabs(n - 50));
-				//para inverter
-				s.Printf(fmt, std::fabs(n - maxScaleAbs));
+                //Print ticks labels
+				if (!m_invert)
+				{
+					s.Printf(fmt, std::fabs(n));
+				}
+				else
+				{
+					s.Printf(fmt, std::fabs(n - maxScaleAbs));
+				}
                 dc.GetTextExtent(s, &tx, &ty);
     #ifdef MATHPLOT_DO_LOGGING
                 if (ty != labelHeigth) wxLogMessage(wxT("mpScaleY::Plot: ty(%f) and labelHeigth(%f) differ!"), ty, labelHeigth);
@@ -3135,6 +3151,19 @@ bool mpWindow::AddLayer( mpLayer* layer, bool refreshDisplay )
     	return true;
     	};
     return false;
+}
+
+double mpWindow::getMinWLayersY()
+{
+	double min = std::numeric_limits<double>::max();
+	for (auto layer : m_layers)
+	{
+		if (layer->GetMinY() < min)
+		{
+			min = layer->GetMinY();
+		}
+	}
+	return min;
 }
 
 bool mpWindow::DelLayer(
@@ -5156,14 +5185,17 @@ void mpXYArea::Plot(wxDC & dc, mpWindow & w)
 				if (first)
 				{
 					first = FALSE;
-					x0 = x1; c0 = c1;
-					// Código Original
-					/*
-					pts.push_back(wxPoint(x0, w.y2p(0)));
-					pts.push_back(wxPoint(x0, c0));
-					*/
-					pts.push_back(wxPoint(x0, w.y2p(GetMaxY())));
-					pts.push_back(wxPoint(x0, w.y2p(GetMaxY()) - c0));
+					x0 = x1; c0 = c1;					
+					if (!m_invert)
+					{
+						pts.push_back(wxPoint(x0, w.y2p(w.getChartAxisMinY())));
+						pts.push_back(wxPoint(x0, c0));
+					}
+					else
+					{
+						pts.push_back(wxPoint(x0, w.y2p(w.getChartAxisMaxY())));
+						pts.push_back(wxPoint(x0, w.y2p(w.getChartAxisMaxY()) - c0));
+					}
 				}
 				bool outUp, outDown;
 				if ((x1 >= startPx) && (x0 <= endPx)) {
@@ -5211,13 +5243,16 @@ void mpXYArea::Plot(wxDC & dc, mpWindow & w)
 								x1 = endPx;
 							}
 						}
-
 						//dc.DrawLine(x0, c0, x1, c1);
 						// Código Original
-						/*
-						pts.push_back(wxPoint(x1, c1));
-						*/
-						pts.push_back(wxPoint(x1, w.y2p(GetMinY()) - c1));
+						if (!m_invert)
+						{
+							pts.push_back(wxPoint(x1, c1));
+						}
+						else
+						{
+							pts.push_back(wxPoint(x1, w.y2p(w.getChartAxisMinY()) - c1));
+						}
 						UpdateViewBoundary(x1, c1);
 
 						if (m_markCorners)
@@ -5230,11 +5265,15 @@ void mpXYArea::Plot(wxDC & dc, mpWindow & w)
 				if (nextX0) { x0 = *nextX0; wxDELETE(nextX0); }
 				if (nextY0) { c0 = *nextY0; wxDELETE(nextY0); }
 			}
-			// Código Original
-			/*
-			pts.push_back(wxPoint(x0, w.y2p(GetMinY())));
-			*/
-			pts.push_back(wxPoint(x0, w.y2p(GetMaxY())));
+			
+			if (!m_invert)
+			{
+				pts.push_back(wxPoint(x0, w.y2p(w.getChartAxisMinY())));
+			}
+			else
+			{
+				pts.push_back(wxPoint(x0, w.y2p(w.getChartAxisMaxY())));
+			}
 			dc.DrawPolygon(pts.size(), pts.data());
 		}
