+++ include/gpSeries.h
@@ -9,8 +9,9 @@ class gpSeries
 {
 protected:
 	wxString m_Label;
-	lineChartLayer *m_Layer = nullptr;
+	lineChartLayer *m_lineLayer = nullptr;
 	barChartLayer *m_barLayer = nullptr;
+	areaChartLayer *m_areaLayer = nullptr;
 	baseXYData m_Data;
 public:
 	bool IsLabel( wxString CompareLabel );
@@ -18,7 +19,6 @@ public:
 	gpSeries( wxString label );
 	//! Destructor
 	virtual ~gpSeries( void );
-	void invertY();
 	void DataPush( double x, double y );
 	void RefreshChart(gpCHART_KIND gpChart_kind, double samplerate,
 		int fftlenght, wxString customXFormula, wxString customYFormula,
@@ -29,10 +29,12 @@ public:
 	void ShowName( bool show );
 	void SetPen( wxPen pen );
 	void SetBrush(wxBrush brush);
-	void DataClear( void );
-	mpLayer* GetLayer( void );
-	xyMultimap_t GetData( void );
+	void DataClear();
+	void invert(bool value);
+	xyMultimap_t GetData();
+	lineChartLayer* getLineChartLayer();
 	barChartLayer* getBarChartLayer();
+	areaChartLayer * getAreaChartLayer();
 };
 
 #endif
